Eliminate bignum calls from `for`-iteration on arrays (#2855)
This is an afterburner on top of #2831, which implemented the `while` optimisation for `for` loops. Even with that the resulting code was littered with `call`s to {`$B_add`, `$@(im)mut_array_size`, `$B_lt`, `$Array.idx_bigint`}, which hide significant cycle counts and contain redundant conditionals. (Also, with future versions of the cycle counting algorithm the `call` instructions will become heavier.) Now we switch away from _expensive_ (that is _checked_) bignum arithmetic to very cheap operations (`i32.shl` and `i32.add`) since the index space of arrays is so limited that it cannot cause overflows into proper heap-allocated bignums. The underlying type of indexing operations remains `Nat`.

I have implemented a handful of primitives that we now use when lowering `for`-to-`while`. In the case of `vals` loop we increment by offset distances of 4 bytes, such that we can simply add the offset to the array pointer and dereference. This saves a bunch of multiplications. We get cycle savings in the ballpark of 7%, which is not too shabby.

## The new primitives

We introduce four new primitive into the IR, which allow to express the use cases of iterating `vals` and `keys` sequences of arrays:

| Name | Purpose | Used in iteration over | Type |
|------|----------|-----------------------|------|
| `GetPastArrayOffset` | obtain the iteration limit | `vals`, `keys` | `[A] -> Nat` |
| `ValidArrayOffset` | determine whether the iteration can continue | `vals`, `keys` | `Nat -> Nat -> Bool` |
| `NextArrayOffset` | advance iteration state | `vals`, `keys` | `Nat -> Nat` |
| `DerefArrayOffset ` | obtain element from array corresponding to the iteration state  | `vals` | `[A] -> Nat -> A` |

`GetPastArrayOffset` and `NextArrayOffset` need a _spacing hint_ which is needed by the backend only to know by which number to increment the iteration state.

The elimination of expensive bignum arithmetic (Wasm calls) grounds on the fact that arrays have less than 2^30 elements. This makes sure that the iteration state is always
- a compact bignum for `keys` iteration
- a `u32` which always has the two lowest bits cleared (and thus appears as a valid compact signed bignum, a.k.a. `Int`) for `vals` iteration.

The iteration state is always a GC compatible object and thus won't cause problems when being heap allocated by the `await`-transformation.

### How these primitives work with iteration state

In the realm of IR all these primitives are strictly adhering to their types.

There is only one case where the iteration state is not a bona fide `Nat`. Namely when _running Wasm and iterating over an array's_ `vals`. In this case the iteration limit (as well as the iteration variable) can technically become a compact "`Nat`" with the _sign bit_ set (so basically a negative compactly-represented `Int`). Here is the scenario how this can happen:

- assume an array with size `1000_000_000` (i.e. `< 2^30`, since no array can occupy more than 4 GB with each element occupying 4 bytes)
- `GetPastArrayOffset` will give `1000000000 << 2` == `0xee6b2800` as `Vanilla`, pointer bit cleared
- `0xee6b2800` is an "illegal" `Nat`, but a legal `Int` with the value -147483648

Since the value is only consumed by the primitives `ValidArrayOffset`, `NextArrayOffset` and `DerefArrayOffset`, which don't care for this quirk, everything is fine. Even if the loop takes part in the `await` transformation (or any other constraint that moves Wasm locals out into heap-allocated data), the `Vanilla` value has no pointer bit set, and thus behaves correctly w.r.t. GC (since GC won't take Motoko's type information into account due to type erasure).

## TODOs
- [x] IR typing rule for `GetPastArrayOffset` still to be filled in
- [x] ~~don't use `2l` as a magic for `1 : Nat`~~ added comment
- [x] verify that the biggest arrays still give compact bignums by `i32.shl 1` (not relevant for `vals`)  — DONE, see subsection above
- [ ] can we use `OtherPrim` (from `lowering`)?
- [x] improve `OtherPrim "array_len"` (suggested for removal, see #2882)
- [ ] find better names for the primitives (without resorting to `Offset`?)

## Further Opportunities
- shift `n_fields` when building `Array`?
- single-`get` variables can be eliminated (if effects not block), by moving the computation formerly before the `set` to replace the `get`
- `for`-loops shouldn't materialise `i32.const 0` on the stack as the result, as units can be readily materialised anywhere (if ever needed)
- [x] lower `DotE(arr, "size")` to `OtherPrim ("array_len", arr)` — see #2882