Pass trap continuation to prim interpreter, make call sites tail recursive
This commit makes calls to prim interpreter (Mo_values.Prim.prim) tail
recursive by passing a "trap" continuation to it.

Previously call sites looked like this:

    try Prim.prim s env v k
    with Invalid_argument s -> trap exp.at "%s" s

This code is not tail recursive, so recursive calls to the interpreter
causes stack overflow.

Interestingly, in the reproducer for #2844, stack overflow manifests
itself as different errors. In most cases it fails as "execution error,
arithmetic overflow", but I've also seen it fail with segfault. I don't
know if this is supposed to happen, or a bug in OCaml RTS.

Fixes #2844