Enable tentative release artefact builds on `master` (#2913)
This enables building of release products on push-events for the `master` branch. It neither sends them to `cachix`, nor attempts to upload them to any release. As previously those activities only happen when a new `git tag` is created.

I took the liberty to streamline the logic a bit, further opportunities exist (commented in PR) but you have to beat me to it.