CI: Add a new workflow that tries to automatically update nix hashes (#2871)
Usually, when we upgrade some dependencies, we have to manually update the hashes. This involves
 * changing the hash to something else (e.g. changing one character)
 * running the build locally
 * observing the error message with the actual hash
 * pasting that
 * commit and push.

The `nix-update` tool can automate that.

This PR makes Github do that on all pushes (even to feature branches), and if needed, push a fix to that branch.

With #2761 adding even more fixed output derivations, this may be more relevant.