CI: Update the ic bump more rarely
now that the replica devs are developing the replica in the open we get
to bump `ic` very often. Most changes are completely irrelevant to us,
but having a daily `niv ic` commit is noisy, and we don’t want to
download `drun` so often.

So let’s treat `ic` like `nixpkgs`: Every week (not day) a bot creats a
PR. If that is green then the new `drun` didn’t break anything. The PR
gets closed, but not merged, automatically. But we’d notice if it
breaks.

When we _do_ want to bump `ic`, we can do it manually, or reopen one of
those PRs.