feat: Implement heartbeat support, docs and test (#3039)
Support system function `heartbeat : () -> async ()`.

Function`heartbeat`, when declared, is called on every Internet Computer subnet *heartbeat*, scheduling an asynchronous call to the `heartbeat` function.
Due to its `async` return type, a heartbeat function may send messages and await results. The result of a heartbeat call, including any trap or thrown error, is ignored. 
The implicit context switch means that the time the heartbeat body is executed may be later than the time the heartbeat was issued by the subnet.

(thanks to Paul Liu for doing most of the implementation work).