bugfix: Issue 3103, confusing eponymous types from separate candid files (#3109)
Fixes #3103
- [x] repro
- [x] fix 

The problem was using global state to map used type definitions in a candid import to their Motoko translations, for subsequent inclusion as type fields in the imported Motoko actor type.

I find this inclusion of type fields problematic, as I expect a candid author to be able to rename/elide equivalent candid type definitions without breaking the compilation of Motoko clients, but that ship has sailed and users (Enzo's QR code generator) for one have started relying on accessing type components of imported actors, so I'm wary of removing this now.

That still leaves once question though: the actor only includes type components that are reference from the body, not the full candid type environment which would be easy to construct and perhaps expected by the user. Is there any reason for this subtlety or should I just clean up the code to add the full environment of the candid file to the actor's type fields (not just the referenced types). I don't think this will break many client, and might be more natural. The only breakage I can envision is if some unreferenced Candid type definition involved Motoko-unsupported float32 types that would now fail to import as Motoko type fields.