Provide basis for obtaining stable variable memory footprint (#3049)
## How it works

### Algorithm

- ~~get controllers list~~ (would need to be its own controller)
- if caller is ~~one of the controllers~~ (or itself) — trap otherwise
  - if actor specifies `system func preupgrade`
    - perform a GC
    - invoke `preupgrade`
  - compute the serialised length of the object containing all stable variables
    (this step will trap if `Nat32` wraps around)
  - convert to `Nat64`
  - add the type descriptor's length (converted to `Nat64`)
  - return it

### Rationale

A GC helps expanding the usable heap so that `preupgrade` won't be hampered by fragmentation.
Before calling the user-supplied `preupgrade` hook, we insert a GC new primitive (`ICPerformGC`, non-user-accessible). This also happens in the upgrade scenario, now.

Invoking `preupgrade` is necessary, otherwise custom logic to fill the stable variables would be ignored.

Getting the _serialised length_ of the stable variables is performed by the highly destructive primitive `ICStableSize` (also non-user-accessible) which leaves the heap in a (potentially) damaged state. No heap reads should be performed after it returns.

The endpoint for getting the stable variable footprint must be a `query`, since it may damage the heap, and needs a state rollback.

A primitive (`stableVarInfo`, user-accessible) is provided to get the query entry point such that the canister can obtain its own stable variable statistics.

`Nat64` is the result type because potentially the serialised stable variables may take more than 4 GB space. This is due to sharing, and the fact that serialisation flattens DAGs to trees and then to a buffer. **Caveat**: Internally all computation currently is done as `u32`, and an assertion makes sure that instead of wrapping a trap occurs on overflow. This means that every call to the primitive `stableVarInfo` should be wrapped in a `try` ... `catch` block.

###  Internal endpoint

The name of the internal query endpoint is `__motoko_stable_var_size`. In the future we should suppress the user-defined endpoints that have a `__` prefix (#3117).

### Lowering
The lowering of the endpoint logic is modelled like `$__get_candid_interface_tmp_hack`.

## TODOs
- [x] test explosion by sharing — e11b6fa58, `query-footprint-overflow.mo` tests when 4GB is exceeded.
- [x] ~~add flag whether to generate footprint query endpoint~~ — Moot!
- [ ] check for name conflicts: added method should not already be part of the interface — deterred to #3117
- [ ] should the query check whether it is called by the [canister's controller](https://forum.dfinity.org/t/how-to-get-the-controller-inside-the-canister-with-rust-motoko/5102)? (see also https://github.com/dfinity/cdk-rs/issues/16#issuecomment-1014052172) — Moot: only controllers can call `canister_status` :-(
- [x] perform a GC before calling the user-defined `preupgrade`?
- [x] test whether the query appears in the Candid (should not!)
    - generated IDL (manually)
    - `$__get_candid_interface_tmp_hack` reply — 476391687fbc
- [x] make the query self-callable (higher-order primitive) — 48df8d1fc
- [x] test the `u32` overflow case
- [ ] we currently drop the count of all references that might be contained in the stable variables (is this a problem?)
- [ ] should we also return other GC-related stats? — Not yet. The result record is extensible.