Inline prim-wrapping functions
The prelude/prim.mo is full of functions simply wrapping a prim, e.g.

   func int64ToNat64(n : Int64) : Nat64 = (prim "num_wrap_Int64_Nat64" : Int64 -> Nat64) n;

generating a Wasm function for them and calling them is absurdly expensive
when the prim is just a simple Wasm instruction. Also, it requires boxing
and unboxing arguments and results.

So we recognize such functions when creating the `const` summary, and
use the prim directly when calling such function.