nix: get releaseVersion from Changelog.md via nix
previously, this only happened when built via the github release action,
but that’s round-about, when we can read and parse `Changelog.md`
directly in nix. Now even `nix-build -A moc` produces a file with a
proper version number.

But it would be a lie to just say `moc-0.6.15` when built that way, as
there are likely changes on top of 0.6.15. So append `-dirty` _unless_
this build is done via the github action and thus likely an official
clean release (the github action even compares git tag with changelog
file).

Drive-by-contributions: use `--omit-metadata` in `test/compare-wat.sh`.

fixes #3167