Make module `Serialization` parametrised in the output streaming direction (#3161)
This is a prelude to #3149. It doesn't change the generated code, just introduces a few holes in the serialiser that can be plugged with code fragments from a `Stream`. Later there will be different `Stream` implementations, but for now we only have the legacy `BumpStream` one. (Actually, there is a minor change in the generated code, which reflects how the stream now `reserve`s a small scratch area for writing, whereas previously it would write right away, and bump the pointer thereafter.)

**Big caveat**: The generated functions coming from `Serialization` (by `Func.share_code`) are still named the same. To make the naming sane, the `Stream` needs to bring its own name plumber. Since we only instantiate once in this PR, this is currently not a problem.

TODO: We definitely need two more interfaces to make this work with generic streams:
- `setup`: create the stream token
- `destroy`: flush the stream and terminate the token.

These are not mandatory in this PR, but when the first other stream appears.