Backend: Unboxed locals
Right now, this code

   func go() {
    let x : Nat32 = 1 +% 1;
    assert (x *% x == 4);
   };

stores the the 32 bit nat in `x` in vanilla representation; doing
complicated shifts and possibly even allocation. To improve this, we use
the type in the AST node to determine if we can use a better stack
representation than `Vanilla`; in this case `UnboxedWord32`, and try to
use that both when filling patterns and when reading variables. This has
the following nice effect:

```diff
   (func $go (type 4) (param $clos i32)
     (local $x i32)
     i32.const 1
     i32.const 1
     i32.add
-    call $box_i32
     local.tee $x
-    call $unbox_i32
     local.get $x
-    call $unbox_i32
     i32.mul
     i32.const 4
     i32.eq
     if  ;; label = @1
     else
       i32.const 65652
       i32.const 51
       call 27
       unreachable
     end)
```