feat: Implement system method `inspect` to satisfy popular demand (#3210)
Implement new system method "inspect" to selectively accept or reject ingress update calls (even of queries called in update mode).

`inspect` takes a record of arguments indicating the `caller`, argument `blob`, and a variant, `msg`, of strongly typed decoding functions. The variant tag indicates the name of the underlying message; its payload is a strongly typed decoder of the (eagerly captured but lazily decoded) blob. Unlike other system methods, that have fixed types, the expected type of `inspect_message` is derived from the type of the actor's fields - in particular, their method arguments.

The argument is a record so we can extend it in future. An example might be providing a new field: the message name as text so we can do content based filtering (think regexp).

Example:

```motoko
actor {

   var c = 0;

   public func inc() : async () { c += 1 };
   public func set(n : Nat) : async () { c := n };
   public query func read() : async Nat { c };
   public func reset() : () { c := 0 }; // oneway

   system func inspect(
     { caller : Principal;
       arg : Blob;
       msg : {
         #inc : Any;
         #set : () -> Nat;
         #read : Any;
         #reset : () -> ();
      }
    }
    ) : Bool  {
      if (arg.size() > 512) return false;
      switch (msg) {
         case (#inc _) { true };
         case (#read _) { true };
         case (#set n)  { (n()) != 666 };
         case (#reset f) { false };
      }
     }
};
```

Full inspection:

[test/run-drun/inspect_message.mo](https://github.com/dfinity/motoko/blob/a844b15b567cc0ff1ac9d661d6ae4d7b9ea1cbf9/test/run-drun/inspect_message.mo)

Selective inspection:
[test/run-drun/inspect_message_blob.mo](https://github.com/dfinity/motoko/blob/a844b15b567cc0ff1ac9d661d6ae4d7b9ea1cbf9/test/run-drun/inspect_message_blob.mo)

Future enhancements: 

It's annoying that the user has to specify the type of the variant if they want to use access it. They can promote some arguments to `Any`, but do have to list the complete superset of expected variants (including hidden `__*` underscore methods. It would be nice if we could just allow the record pattern to be unannotated by using the expected type, in analysis mode. 
But let's see how many user's actually use this optional feature before we invest more effort into it.