doc: markdown docs (#3172)
# Convert adoc to md and incorporate base library docs

In this PR, the markdown docs are generated from the adoc sources into `doc/md`, `docs/md/examples` and `docs/md/base`
using Makefile targets `make base` and `make md` (which should be run in a nix-shell).

The generated files are checked in, not a build product, to ease incorporation from the DFINITY portal.

The manual adoc rewrite and automatic translation:

- adjusts the adoc to only use features supported by conversion (GH reference and labels syntax, simple tables)
- runs mo-doc to produce markdown documentation including additional file index.md
- rewrites links
- rewrite adoc admonitions to mdx admonitions used by the portal
- supports file inclusion of entire files or specific line ranges (but not tags) use remark-code-import plugin by rewriting the adoc includes (asciidoctor otherwise inlines the include code) so that included files can be checked separately by mo-doc
- preserves adoc block comments (but not line comments) for TODO etc
- rewrites adoc interpreter code attributes to MDX ones
- content fixes spotted during the transition.

The best way to preview the documentation is to open a PR on the portal, and point the motoko git submodule to the branch of your PR and wait for CI to produce a preview.

The next step is to remove the adoc sources and rewrite/remove any related nix derivations, leaving only the markdown as the source of truth, but that can be done in a separate PR.