Backend: get_aliased_box should not create boxes
the semantics of the initializer of a mutable field in NewObjE is that
it should alias the value; this is needed for async/await.

To be aliased, that variable must be put in the heap, as a HeapInd.
Previously, if that was not the case, we’d simply create a new box. But
that breaks the semantics! (I don’t think this could be observed in
source, though).

Better if get_aliased_box fails if it doesn’t find the value allocated
on the heap. And to make sure it always is on the heap, we have to treat
the initializer in NewObjE a bit like a mutable variable captured by a
lambda. And then it seems to work.