alt syntax for record extension (#3367)
This PR switches the order of fields with bases and adds the new `with` keyword to separate them:
``` Motoko
// before
{ f1 = v1; ... in base1 and base2 and ... }
// after
{ base1 and base2 and ... with f1 = v1; ...; }
```
---------

variant of #3084 with alternative syntax

note this only changes the syntax, not evaluation order, which would need to be adjusted in both the interpreter and desugaring to evaluate bases before fields. The order of the arguments to ObjE should also be swapped.