Make overwriting of `var` fields from bases obligatory (#3411)
Make mutable field aliasing (when extending/merging records) an experimental opt-in feature, and hide it behind a feature flag. When the flag is not supplied, an error is emitted that instructs the user to overwrite all `var` fields from bases explicitly. When `--experimental-field-aliasing` is given no such error gets emitted, and the mutable cells of bases are aliased in the resulting record.

See discussion in #3084 why this is a necessity.

Also contains a few minor tweaks.

Co-authored-by: Claudio Russo <claudio@dfinity.org>