Backend: Correctly join Const and UnboxedTuple reps (#3551)
previously the code assumed that we never join a constant value and a unboxed tuple. If it tried, it would print a warning, and simply use one of the two.

@crusso reports that this was observed in the wild, and indeed such joins are easy to trigger.

I don’t recall why I thought that this combination could not occur. In fact, it seems that the best stackrep when a const value meets a non-const value is to use that other value's stack representation. So lets put that in.