Multi-value emulation: Use globals, not allocated tuples
So far, the IC does not support wasm using multi-value returns from
functions and blocks (see #1459). The backend was originally written
with that feature in mind, and later we added a work-around (module
`FakeMultiVal`).

For functions we work around it by storing multiple values in fixed
globals, and reading them after the call. No allocations needed.

For blocks with multiple values, however, we allocate a tuple (array) on the
heap, write the values there, and read them later. This adds unnecessary
heap churn and fills up memory faster.

So this PR uses the mechanism from the functions also for blocks.

Goes well with #3556