Multi-value emulation: Use globals, not allocated tuples (#3566)
So far, the IC does not support wasm using multi-value returns from functions and blocks (see #1459). The backend was originally written with that feature in mind, and later we added a work-around (module `FakeMultiVal`).

For functions we work around it by storing multiple values in fixed globals, and reading them after the call. No allocations needed.

For blocks with multiple values, however, we allocate a tuple (array) on the heap, write the values there, and read them later. This adds unnecessary heap churn and fills up memory faster.

So this PR uses the mechanism from the functions also for blocks.

Goes well with #3556

Here is a (particularly bad) example:

In
```motoko
func multi_value_blocks(n : Nat) : (Nat, Nat) {
  return (
    if (n == 0) {
      returns_tuple()
    } else {
      if (n == 1) {
        returns_tuple()
      } else {
        returns_tuple()
      }
   }
 );
};
```
Now no allocation happens (and all the reads and writes to the global variables are eliminated due to #3556):

```diff
   (func $multi_value_blocks (type 1) (param $clos i32) (param $n i32)
     local.get $n
     i32.const 0
     call $B_eq
-    if (result i32)  ;; label = @1
+    if  ;; label = @1
       i32.const 0
       call $returns_tuple
-      global.get 4
-      global.get 5
-      call $to_2_tuple
     else
       local.get $n
       i32.const 2
       call $B_eq
-      if (result i32)  ;; label = @2
+      if  ;; label = @2
         i32.const 0
         call $returns_tuple
-        global.get 4
-        global.get 5
-        call $to_2_tuple
       else
         i32.const 0
         call $returns_tuple
-        global.get 4
-        global.get 5
-        call $to_2_tuple
       end
     end
-    call $from_2_tuple
     return)
```