Add source locations and exported types to `moc.js` parser output  (#3455)
The main addition in this PR is `Mo_def.Arrange_source`, a more verbose alternative to `Mo_def.Arrange` which includes source locations in the generated s-expression tree. This will be used in the new `dfx`-agnostic VS Code language server as well as in a future upgrade to `mo-doc`. 

Follow-up questions:
- Is there a previously existing way to intersperse type information throughout an AST? 
  - Anything similar to transforming `1 + 2` to `((1: Nat) + (2: Nat)): Nat` would work for the anticipated use cases. 
- Does anything currently rely on the `Mo_def.Arrange` s-expressions, aside from debug output? Since there's a lot of duplicated code between `Arrange` and `Arrange_source`, it probably makes sense to either combine the modules (and pass along a config value), or simply deprecate `Arrange` and include source information everywhere. Curious to hear any opinions about this.