experiment: implement basic trampoline for moc ast interpreter (#3541)
Although we support browser based execution of the moc interpreter, it currently only runs on tiny examples before exhausting the JS stack. 

This is because the interpreter is written in CPS and ocaml_js does not honour tail calls to unknown functions, i.e. every call of a (dynamic) continuation. 

This PR abuses our existing event loop to implement a simple trampoline for the cps based interpreter, exiting the current interpretation to clear the stack, but resuming it from the front of the event queue.

This allows the js compiled interpreter to run in the browser without running out of stack.

TODO: 
- [x] only do this JS code, not OCaml native code.
- [x] optimize to avoid allocation
- [ ] use a stack local counter, or hide it in the environment?
- [ ] use a dequeue instead of abusing the queue (for more efficient queue jumping)?
- [x] interpretation of prim expressions uses try match and isn't tail rec. FIX FIRST, probably by merging https://github.com/dfinity/motoko/pull/2895

Now, instead of sorting a buffer of 4 elements failing with stack overflow, we can sort hundreds of elements!

Screenshot: 

Before:
![image](https://user-images.githubusercontent.com/85788/199482032-6c1024a0-83c5-494a-8dee-182320322bc7.png)

After:

![image](https://user-images.githubusercontent.com/85788/199481478-06021e6a-9dfd-4865-9c9e-21b9ac6a7ba7.png)