experiment: Viper integration (#3477)
# Motoko-san, an experimental verification method

This PR adds (very experimental) Viper code generation from (a limited subset of) Motoko. This is a proof of concept, showing that we can generate appropriate assertions so that Viper can diagnose concurrency (e.g. re-entrancy) bugs. A few tests are provided to demonstrate this.

Note that this is (a.t.m.) _not_ directed towards the general public, rather towards folks with verification experience.

To use the Viper capabilities
- specify `--viper` on the command line (Viper generation)
- set `env MOC_UNLOCK_VERIFICATION=1` (regular compilation using extended grammar)

When the verification features are enabled, a convenience logical operator `implies` becomes a keyword. It has the following truth table:

| a | b | a `implies` b |
|-|-|-|
| false | false | true |
| false | true | true |
| true | false | false |
| true | true | true |