bugfix: fix spurious discarded field errors in --stable-compatible check (#3645)
Fix bug in stable-compatiblity check reported here:

https://forum.dfinity.org/t/weird-warning-message-when-deploying-existing-canister/17276/5.

The problem is that the verbose check (that emits multiple diagnostic errors) would, on finding a discarded field from the lhs,  recurse only on the tail of the rhs, not the rhs itself, causing all remaining fields of the lhs to be reported as discarded, even when present in the (original) rhs.

Added a small repro and manually verified it fixes the issue on the full `*.most` files shared by the user in a private forum message.

The bug itself is benign though obviously confusing to the user.