make `setGlobalTimer` a scoped capability (#3614)
In the review walkthrough (SGM 2022-11-29) @crusso suggested that `setGlobalTimer` should be a capability in the scope of `system func timer` only.

This shields libraries from moving the global timer's expiration around without central coordination.

There is an initial call to the `system func timer` just after the canister is started, to allow setting up the first actual expiration. This initial priming of the global timer is only added when there is a _user-written_ `system func timer` present, and internally we detect this in the backend by observing a source location on the function body. (Defaulted and disabled timers will have no non-trivial srcloc on the body.)

_TODO_: deactivated tests (compacting GC in `ic-ref` are a possible indication of GC bug!)

Co-authored-by: Nix hash updater <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Claudio Russo <claudio@dfinity.org>