Support private functions and `old` expressions in Motoko-to-Viper translation (#3675)
This PR adds support for translating private Motoko functions into Viper. The translation is mostly analogous to that for public `async` functions. However, a private function does not have to be called from a state that satisfies the canister invariant (i.e., it may be called even in intermediate states), and a private function does not have to preserve the canister invariant (i.e., it may result in an intermediate state).

Additionally, we add `old` expressions to enable useful specs in private actor functions, e.g.:

```motoko
private func reward() : () {
  count += 1;
  assert:return count == old count + 1;
};
```